# perf: optimize point tripling