--- a/ecc/bls12-381/g2.go
+++ b/ecc/bls12-381/g2.go
@@ -694,12 +694,12 @@ func (p *G2Jac) mulGLV(q *G2Jac, s *big.Int) *G2Jac {
 	table[6].Set(&table[3]).AddAssign(&table[2])
 	table[7].Double(&table[3])
 	table[8].Set(&table[7]).AddAssign(&table[0])
-	table[9].Set(&table[7]).AddAssign(&table[1])
+	table[9].Double(&table[4])
 	table[10].Set(&table[7]).AddAssign(&table[2])
 	table[11].Triple(&table[3])
 	table[12].Set(&table[11]).AddAssign(&table[0])
 	table[13].Set(&table[11]).AddAssign(&table[1])
-	table[14].Set(&table[11]).AddAssign(&table[2])
+	table[14].Triple(&table[4])
 
 	// bounds on the lattice base vectors guarantee that k1, k2 are len(r)/2 or len(r)/2+1 bits long max
 	// this is because we use a probabilistic scalar decomposition that replaces a division by a right-shift