# Description

Optimize G1Jac tripling (20% faster for bn254, 10% for bls12-377).


```
benchmark                   old ns/op     new ns/op     delta
BenchmarkG1JacTriple-32     290           233           -19.49%
BenchmarkG2JacTriple-32     546           542           -0.73%
```